backend/requirements.txt
bcrypt==4.1.3
groq==0.28.0
httpx==0.27.2
numpy==1.26.4
PyPDF2==3.0.1
python-dotenv==1.0.1
PyMuPDF==1.24.5
flask==3.1.3
flask-cors==6.0.2
PyJWT==2.12.1
gunicorn==22.0.0
faiss-cpu==1.12.0
sentence-transformers==3.0.1
rank-bm25==0.2.2
hypothesis==6.98.3
